Moved grid making to own file, updated examples, bug fixes
Created make_grids.py to house all grid making code for easier reference
